+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-09-24
+### Added
+- Implemented the Spot API in `HTTP`
+- Implemented the Account Asset API in `HTTP`
+- Implemented the Spot websocket in `WebSocket`
+
 ## [1.2.1] - 2021-07-15
-### Remoed
+### Removed
 - Removed `position_mode_switch()` as the endpoint has not been released to mainnet yet
 
 ## [1.2.0] - 2021-07-09